requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
addict=2.4.0=pypi_0
argon2-cffi=20.1.0=pypi_0
asttokens=2.0.4=pypi_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
attrs=20.3.0=pypi_0
auto-tune=0.1.0=pypi_0
backcall=0.2.0=pypi_0
bleach=3.3.0=pypi_0
bypy=1.6.11=pypi_0
ca-certificates=2021.1.19=h06a4308_0
cached-property=1.5.2=pypi_0
captcha=0.3=pypi_0
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.4=pypi_0
cftime=1.4.1=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
cmake=3.18.4.post1=pypi_0
coco=0.0.0=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.22=pypi_0
decorator=4.4.2=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.3=pypi_0
easydict=1.9=pypi_0
entrypoints=0.3=pypi_0
flask=1.1.4=pypi_0
flask-cors=3.0.10=pypi_0
future=0.18.2=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.36.0=pypi_0
gunicorn=20.1.0=pypi_0
h5py=3.3.0=pypi_0
hccl=0.1.0=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=3.7.2=pypi_0
iniconfig=1.1.1=pypi_0
ipykernel=5.5.3=pypi_0
ipython=7.22.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.3=pypi_0
itsdangerous=1.1.0=pypi_0
jedi=0.18.0=pypi_0
jieba=0.42.1=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.0.1=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.12=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter-contrib-core=0.3.3=pypi_0
jupyter-core=4.7.1=pypi_0
jupyter-highlight-selected-word=0.2.0=pypi_0
jupyter-latex-envs=1.4.6=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
kiwisolver=1.3.1=pypi_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
lxml=4.6.3=pypi_0
markupsafe=1.1.1=pypi_0
marshmallow=3.13.0=pypi_0
matplotlib=3.4.1=pypi_0
mindinsight=1.3.0=pypi_0
mindspore-ascend=1.2.0=pypi_0
mistune=0.8.4=pypi_0
mmcv=0.2.14=pypi_0
mpmath=1.1.0=pypi_0
multiprocess=0.70.11.1=pypi_0
nbclient=0.5.3=pypi_0
nbconvert=5.6.1=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
netcdf4=1.5.5=pypi_0
networkx=2.5.1=pypi_0
nltk=3.6.2=pypi_0
notebook=6.3.0=pypi_0
numpy=1.17.5=pypi_0
opencv-contrib-python=4.5.3.56=pypi_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1i=h27cfd23_0
packaging=20.9=pypi_0
pandas=1.2.4=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.8.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.1.0=pypi_0
pip=20.3.3=py37h06a4308_0
pluggy=0.13.1=pypi_0
prettytable=2.1.0=pypi_0
prometheus-client=0.10.1=pypi_0
prompt-toolkit=3.0.18=pypi_0
protobuf=3.14.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.7.0=pypi_0
py=1.10.0=pypi_0
pycocotools=2.0.2=pypi_0
pycparser=2.20=pypi_0
pygments=2.8.1=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.17.3=pypi_0
pytest=6.2.2=pypi_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.0.3=pypi_0
qtconsole=5.0.3=pypi_0
qtpy=1.9.0=pypi_0
readline=7.0=h7b6447c_5
regex=2021.4.4=pypi_0
requests=2.25.1=pypi_0
requests-toolbelt=0.9.1=pypi_0
rouge=1.0.0=pypi_0
sacremoses=0.0.45=pypi_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.0=pypi_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pypi_0
setuptools=52.0.0=py37h06a4308_0
shapely=1.7.1=pypi_0
simpleitk=2.0.2=pypi_0
six=1.15.0=pypi_0
sqlite=3.33.0=h62c20be_0
subword-nmt=0.3.7=pypi_0
sympy=1.7.1=pypi_0
te=0.4.0=pypi_0
terminado=0.9.4=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2021.4.8=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.2=pypi_0
topi=0.4.0=pypi_0
torch=1.8.1+cpu=pypi_0
torchaudio=0.8.1=pypi_0
torchvision=0.9.1+cpu=pypi_0
tornado=6.1=pypi_0
tqdm=4.59.0=pypi_0
traitlets=5.0.5=pypi_0
treelib=1.6.1=pypi_0
typing-extensions=3.7.4.3=pypi_0
unknown=0.0.0=pypi_0
urllib3=1.26.4=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xlsxwriter=1.4.4=pypi_0
xmltodict=0.12.0=pypi_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yapf=0.31.0=pypi_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
